Last to int->float64 conversions for API safety

+++ api.go
@@ -37,8 +37,8 @@ type WeatherResponse struct {
 	Base string `json:"base"`
 	Main struct {
 		Temp     float64 `json:"temp"`
-		Pressure int     `json:"pressure"`
-		Humidity int     `json:"humidity"`
+		Pressure float64 `json:"pressure"`
+		Humidity float64 `json:"humidity"`
 		TempMin  float64 `json:"temp_min"`
 		TempMax  float64 `json:"temp_max"`
 	} `json:"main"`
